move immutable tests out of test_tool.py and rename files
